voxera-os-scaffold/voxera-os/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voxera-os"
version = "0.1.0"
description = "Voxera OS scaffold: voice-first AI control plane on Linux."
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
authors = [{name="Voxera OS Contributors"}]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0",
  "pydantic>=2.6",
  "httpx>=0.27",
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "jinja2>=3.1",
  "keyring>=25.0",
  "platformdirs>=4.2",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.6",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "mypy>=1.10",
  "types-PyYAML>=6.0",
  "pre-commit>=3.7",
]

[project.scripts]
voxera = "voxera.cli:app"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
